Use the targetset to identify a group of targets
